move: to models package + use IMessageSignature
